backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.6.3
asttokens==2.4.1
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cohere==5.3.2
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.108.0
fastavro==1.9.4
filelock==3.13.4
fsspec==2024.3.1
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.1
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
httpx-sse==0.4.0
huggingface-hub==0.22.2
idna==3.6
ipython==8.18.1
jedi==0.19.1
json5==0.9.25
matplotlib-inline==0.1.6
numpy==1.26.3
openai==1.17.1
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pocketbase==0.10.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==13.13
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.32.0.post1
tokenizers==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-requests==2.31.0.20240406
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
vertexai==0.0.1
wcwidth==0.2.12
